+++ supervision/detection/annotate.py
@@ -3,16 +3,12 @@ from typing import List, Optional, Union
 import cv2
 
 from supervision.annotators.base import ImageType
-from supervision.annotators.utils import scene_to_annotator_img_type
 from supervision.detection.core import Detections
 from supervision.draw.color import Color, ColorPalette
+from supervision.utils.conversion import convert_for_annotation_method
 from supervision.utils.internal import deprecated
 
 
-@deprecated(
-    "`BoxAnnotator` is deprecated and will be removed in "
-    "`supervision-0.22.0`. Use `BoundingBoxAnnotator` and `LabelAnnotator` instead"
-)
 class BoxAnnotator:
     """
     A class for drawing bounding boxes on an image using detections provided.
@@ -46,7 +42,11 @@ class BoxAnnotator:
         self.text_thickness: int = text_thickness
         self.text_padding: int = text_padding
 
-    @scene_to_annotator_img_type
+    @deprecated(
+        "`BoxAnnotator` is deprecated and will be removed in "
+        "`supervision-0.22.0`. Use `BoundingBoxAnnotator` and `LabelAnnotator` instead"
+    )
+    @convert_for_annotation_method
     def annotate(
         self,
         scene: ImageType,